migration/stockAndPick/web_app/requirements.txt
# Flask Web Application Dependencies for Stock and Pick PCB Inventory

# Core Flask framework
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
Flask-Caching==2.1.0
Flask-Compress==1.14

# Forms and validation
WTForms==3.0.1

# PostgreSQL database adapter
psycopg2-binary==2.9.7

# Password hashing
bcrypt==4.0.1

# Additional utilities
python-dotenv==1.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
blinker==1.6.3
itsdangerous==2.1.2

# Production WSGI server for better performance
gunicorn==21.2.0

# Access database connectivity (using subprocess with mdb-tools)
# pandas and numpy removed to avoid compatibility issues

# Development dependencies (optional)
pytest==7.4.2
pytest-flask==1.2.0
coverage==7.3.2